refactor: 에러 관련 폴더 구조 재배치
